requirements.txt
aiogram==3.1.1
APScheduler==3.10.4
python-dotenv==1.0.0
Pyrogram==2.0.106
SQLAlchemy==2.0.23